Separated functionality to new file, exported more from index
